+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the create-aptos-dapp tool will be captured in this file.
 # Unreleased
 
 - Add new view function to minting template that returns user mint balance and display that on UI
+- Add a minting enabled / disabled flag to nft minting template and token minting template
 
 # 0.0.24 (2024-09-04)